Refactored Navigation/LinkRow to class component
Reason:

I'm working on a dynamic animation changer for clicking between links.  Currently, the animation of accelerate doesn't perform well for transitions less than 1000px or so.  The linear animation, however, performs more ideal for the smaller transitions.  For this reason, I'm going to work on a way to calculate the current distance to target that will change the animation dynamically.