Restructure Jumbotron for mobile views
In order to start appropriately laying out the Resume component, I have gone back and added style extensions for mobile/landscape and mobile/portrait.

This highlights a larger issue with the position related scrolling this app feature when buttons are clicked.  The numbers have to be adjusted any time that the size of components change.