Toggle in/out buttons working
However, there is an issue with the animations whenever we go to click a tab.  The issue stems from the fact that we are using 2 completely separate divs to accomplish our layers.

To layer one div over the other, I had to increase it's z-index by 1.  This puts it ahead of the other on the Z axis.  However, when a tab is clicked, the structure of the divs were still remaining in place, effectively causing any content below it to be masked over and just not work.  Changing the Z index back to 0 causes the component animations to just disappear.

As of yet, no idea how to fix this and get this component to work the way that I would like it to.